# Conflicts:
#	.travis.yml
#	src/Makefile.in
#	src/info/infoBackup.c
#	test/lib/pgBackRestTest/Common/CiTest.pm
#	test/travis.pl

+++ test/test.pl
@@ -46,7 +46,6 @@ use pgBackRestTest::Common::BuildTest;
 use pgBackRestTest::Common::CodeCountTest;
 use pgBackRestTest::Common::ContainerTest;
 use pgBackRestTest::Common::CoverageTest;
-use pgBackRestTest::Common::CiTest;
 use pgBackRestTest::Common::DefineTest;
 use pgBackRestTest::Common::ExecuteTest;
 use pgBackRestTest::Common::HostGroupTest;
@@ -81,12 +80,13 @@ test.pl [options]
    --build-max          max processes to use for builds (default 4)
    --coverage-only      only run coverage tests (as a subset of selected tests)
    --c-only             only run C tests
+   --container-only     only run tests that must be run in a container
    --gen-only           only run auto-generation
+   --gen-libc           generate libc code to embed in the binary
    --no-gen             do not run code generation
    --code-count         generate code counts
    --smart              perform libc/package builds only when source timestamps have changed
    --no-package         do not build packages
-   --no-ci-config       don't overwrite the current continuous integration config
    --dev                --smart --no-package --no-optimize
    --dev-test           --no-package
    --expect             --no-package --vm=co7 --db=9.6 --log-force
@@ -151,12 +151,13 @@ my $bCoverageOnly = false;
 my $bCoverageSummary = false;
 my $bNoCoverage = false;
 my $bCOnly = false;
+my $bContainerOnly = false;
 my $bGenOnly = false;
+my $bGenLibC = false;
 my $bNoGen = false;
 my $bCodeCount = false;
 my $bSmart = false;
 my $bNoPackage = false;
-my $bNoCiConfig = false;
 my $bDev = false;
 my $bDevTest = false;
 my $bBackTrace = false;
@@ -195,12 +196,13 @@ GetOptions ('q|quiet' => \$bQuiet,
             'build-only' => \$bBuildOnly,
             'build-max=s' => \$iBuildMax,
             'no-package' => \$bNoPackage,
-            'no-ci-config' => \$bNoCiConfig,
             'coverage-only' => \$bCoverageOnly,
             'coverage-summary' => \$bCoverageSummary,
             'no-coverage' => \$bNoCoverage,
             'c-only' => \$bCOnly,
+            'container-only' => \$bContainerOnly,
             'gen-only' => \$bGenOnly,
+            'gen-libc' => \$bGenLibC,
             'no-gen' => \$bNoGen,
             'code-count' => \$bCodeCount,
             'smart' => \$bSmart,
@@ -551,7 +553,7 @@ eval
 
             # Auto-generate C library code to embed in the binary
             #-----------------------------------------------------------------------------------------------------------------------
-            if (!$bSmart || grep(/^libc\//, @stryModifiedList))
+            if ($bGenLibC && (!$bSmart || grep(/^libc\//, @stryModifiedList)))
             {
                 my $strLibC = executeTest(
                     "cd ${strBackRestBase}/libc && " .
@@ -638,13 +640,6 @@ eval
             }
         }
 
-        # Build CI config
-        #---------------------------------------------------------------------------------------------------------------------------
-        if (!$bNoCiConfig)
-        {
-            (new pgBackRestTest::Common::CiTest($oStorageBackRest))->process();
-        }
-
         # Check Perl version against release notes and update version in C code if needed
         #---------------------------------------------------------------------------------------------------------------------------
         my $bVersionDev = true;
@@ -711,23 +706,24 @@ eval
         my $iTestFail = 0;
         my $iTestRetry = 0;
         my $oyProcess = [];
-        my $strCoveragePath = "${strTestPath}/cover_db";
         my $strCodePath = "${strBackRestBase}/test/.vagrant/code";
 
         if (!$bDryRun || $bVmOut)
         {
-            &log(INFO, "cleanup old data and containers");
-            containerRemove('test-([0-9]+|build)');
+            &log(INFO, "cleanup old data" . ($strVm ne VM_NONE ? " and containers" : ''));
+
+            if ($strVm ne VM_NONE)
+            {
+                containerRemove('test-([0-9]+|build)');
+            }
 
             for (my $iVmIdx = 0; $iVmIdx < 8; $iVmIdx++)
             {
                 push(@{$oyProcess}, undef);
             }
 
-            executeTest(
-                "sudo rm -rf ${strTestPath}/cover_db ${strTestPath}/test-* ${strTestPath}/expect-*" .
-                ($bDev ? '' : " ${strTestPath}/gcov-*"));
-            $oStorageTest->pathCreate($strCoveragePath, {strMode => '0770', bIgnoreExists => true, bCreateParent => true});
+            executeTest("rm -rf ${strTestPath}/test-* ${strTestPath}/data-*" . ($bDev ? '' : " ${strTestPath}/gcov-*"));
+            $oStorageTest->pathCreate($strTestPath, {strMode => '0770', bIgnoreExists => true, bCreateParent => true});
 
             # Remove old coverage dirs -- do it this way so the dirs stay open in finder/explorer, etc.
             executeTest("rm -rf ${strBackRestBase}/test/coverage/c/*");
@@ -760,7 +756,7 @@ eval
         {
             # Get the test list
             $oyTestRun = testListGet(
-                $strVm, \@stryModule, \@stryModuleTest, \@iyModuleTestRun, $strPgVersion, $bCoverageOnly, $bCOnly);
+                $strVm, \@stryModule, \@stryModuleTest, \@iyModuleTestRun, $strPgVersion, $bCoverageOnly, $bCOnly, $bContainerOnly);
 
             # Determine if the C binary and test library need to be built
             foreach my $hTest (@{$oyTestRun})
@@ -869,10 +865,13 @@ eval
                     {
                         &log(INFO, "    build bin for ${strBuildVM} (${strBuildPath})");
 
-                        executeTest(
-                            "docker run -itd -h test-build --name=test-build" .
-                            " -v ${strBackRestBase}:${strBackRestBase} " . containerRepo() . ":${strBuildVM}-build",
-                            {bSuppressStdErr => true});
+                        if ($strBuildVM ne VM_NONE)
+                        {
+                            executeTest(
+                                "docker run -itd -h test-build --name=test-build" .
+                                " -v ${strBackRestBase}:${strBackRestBase} " . containerRepo() . ":${strBuildVM}-build",
+                                {bSuppressStdErr => true});
+                        }
 
                         foreach my $strBinSrcPath (@stryBinSrcPath)
                         {
@@ -889,17 +888,21 @@ eval
                         if ($bBuildOptionsDiffer || !$oStorageBackRest->exists("${strBuildPath}/Makefile"))
                         {
                             executeTest(
-                                "docker exec -i test-build bash -c 'cd ${strBuildPath} && ./configure${strConfigOptions}'",
+                                ($strBuildVM ne VM_NONE ? 'docker exec -i test-build ' : '') .
+                                "bash -c 'cd ${strBuildPath} && ./configure${strConfigOptions}'",
                                 {bShowOutputAsync => $bLogDetail});
                         }
 
                         executeTest(
-                            'docker exec -i test-build' .
-                            " make -j ${iBuildMax}" . ($bLogDetail ? '' : ' --silent') .
+                            ($strBuildVM ne VM_NONE ? 'docker exec -i test-build ' : '') .
+                            "make -j ${iBuildMax}" . ($bLogDetail ? '' : ' --silent') .
                                 " --directory ${strBuildPath} CFLAGS='${strCFlags}' LDFLAGS='${strLdFlags}'",
                             {bShowOutputAsync => $bLogDetail});
 
-                        executeTest("docker rm -f test-build");
+                        if ($strBuildVM ne VM_NONE)
+                        {
+                            executeTest("docker rm -f test-build");
+                        }
                     }
                 }
             }
@@ -926,7 +929,7 @@ eval
                 foreach my $strBuildVM (@stryBuildVm)
                 {
                     my $strBuildPath = "${strLibCPath}/${strBuildVM}/libc";
-                    my $bContainerExists = $strBuildVM ne $strVmHost;
+                    my $bContainerExists = $strBuildVM ne $strVmHost && $strBuildVM ne VM_NONE;
                     my $strConfigOptions = (vmDebugIntegration($strBuildVM) ? ' --enable-test' : '');
 
                     my $strLibCSmart = "${strBuildPath}/blib/arch/auto/pgBackRest/LibC/LibC.so";
@@ -942,12 +945,6 @@ eval
 
                             $bRebuild = true;
                         }
-                    }
-
-                    # Delete old libc files from the host
-                    if ($bRebuild)
-                    {
-                        executeTest('sudo rm -rf ' . $oVm->{$strBuildVM}{&VMDEF_PERL_ARCH_PATH} . '/auto/pgBackRest/LibC');
                     }
 
                     if ($bRebuild)
@@ -1014,19 +1011,13 @@ eval
                         {
                             executeTest("docker rm -f test-build");
                         }
-
-                        if ($strBuildVM eq $strVmHost)
-                        {
-                            executeTest("sudo make -C ${strBuildPath} install", {bSuppressStdErr => true});
-                            buildLoadLibC();
-                        }
                     }
                 }
             }
 
             # Build the package
             #-----------------------------------------------------------------------------------------------------------------------
-            if (!$bNoPackage)
+            if (!$bNoPackage && $strVm ne VM_NONE)
             {
                 my $strPackagePath = "${strVagrantPath}/package";
                 my $strPackageSmart = "${strPackagePath}/build.timestamp";
@@ -1055,7 +1046,7 @@ eval
                         &log(INFO, 'package dependencies have changed, rebuilding...');
                     }
 
-                    executeTest("sudo rm -rf ${strPackagePath}");
+                    executeTest("rm -rf ${strPackagePath}");
                 }
 
                 # Loop through VMs to do the package builds
@@ -1070,16 +1061,19 @@ eval
                     {
                         &log(INFO, "build package for ${strBuildVM} (${strBuildPath})");
 
-                        executeTest(
-                            "docker run -itd -h test-build --name=test-build" .
-                            " -v ${strBackRestBase}:${strBackRestBase} " . containerRepo() . ":${strBuildVM}-build",
-                            {bSuppressStdErr => true});
+                        if ($strVm ne VM_NONE)
+                        {
+                            executeTest(
+                                "docker run -itd -h test-build --name=test-build" .
+                                " -v ${strBackRestBase}:${strBackRestBase} " . containerRepo() . ":${strBuildVM}-build",
+                                {bSuppressStdErr => true});
+                        }
 
                         $oStorageBackRest->pathCreate($strBuildPath, {bIgnoreExists => true, bCreateParent => true});
 
                         executeTest("rsync -r --exclude .vagrant --exclude .git ${strBackRestBase}/ ${strBuildPath}/");
                         executeTest(
-                            "docker exec -i test-build " .
+                            ($strVm ne VM_NONE ? "docker exec -i test-build " : '') .
                             "bash -c 'cp -r /root/package-src/debian ${strBuildPath}' && sudo chown -R " . TEST_USER .
                             " ${strBuildPath}");
 
@@ -1123,15 +1117,18 @@ eval
                             ${$oStorageBackRest->get("${strBuildPath}/debian/changelog")});
 
                         executeTest(
-                            "docker exec -i test-build " .
+                            ($strVm ne VM_NONE ? "docker exec -i test-build " : '') .
                             "bash -c 'cd ${strBuildPath} && debuild -i -us -uc -b'");
 
                         executeTest(
-                            "docker exec -i test-build " .
+                            ($strVm ne VM_NONE ? "docker exec -i test-build " : '') .
                             "bash -c 'rm -f ${strPackagePath}/${strBuildVM}/*.build ${strPackagePath}/${strBuildVM}/*.changes" .
                             " ${strPackagePath}/${strBuildVM}/pgbackrest-doc*'");
 
-                        executeTest("docker rm -f test-build");
+                        if ($strVm ne VM_NONE)
+                        {
+                            executeTest("docker rm -f test-build");
+                        }
                     }
 
                     if (!$oStorageBackRest->pathExists($strBuildPath) && $oVm->{$strBuildVM}{&VM_OS_BASE} eq VM_OS_BASE_RHEL)
@@ -1139,10 +1136,13 @@ eval
                         &log(INFO, "build package for ${strBuildVM} (${strBuildPath})");
 
                         # Create build container
-                        executeTest(
-                            "docker run -itd -h test-build --name=test-build" .
-                            " -v ${strBackRestBase}:${strBackRestBase} " . containerRepo() . ":${strBuildVM}-build",
-                            {bSuppressStdErr => true});
+                        if ($strVm ne VM_NONE)
+                        {
+                            executeTest(
+                                "docker run -itd -h test-build --name=test-build" .
+                                " -v ${strBackRestBase}:${strBackRestBase} " . containerRepo() . ":${strBuildVM}-build",
+                                {bSuppressStdErr => true});
+                        }
 
                         # Create build directories
                         $oStorageBackRest->pathCreate($strBuildPath, {bIgnoreExists => true, bCreateParent => true});
@@ -1152,7 +1152,9 @@ eval
                         $oStorageBackRest->pathCreate("${strBuildPath}/BUILD", {bIgnoreExists => true, bCreateParent => true});
 
                         # Install PostreSQL 11 development for package builds
-                        executeTest("docker exec -i test-build bash -c 'yum install -y postgresql11-devel 2>&1'");
+                        executeTest(
+                            ($strVm ne VM_NONE ? "docker exec -i test-build " : '') .
+                            "bash -c 'yum install -y postgresql11-devel 2>&1'");
 
                         # Copy source files
                         executeTest(
@@ -1162,7 +1164,7 @@ eval
 
                         # Copy package files
                         executeTest(
-                            "docker exec -i test-build bash -c '" .
+                            ($strVm ne VM_NONE ? "docker exec -i test-build " : '') . "bash -c '" .
                             "ln -s ${strBuildPath} /root/rpmbuild && " .
                             "cp /root/package-src/pgbackrest.spec ${strBuildPath}/SPECS && " .
                             "cp /root/package-src/*.patch ${strBuildPath}/SOURCES && " .
@@ -1190,12 +1192,15 @@ eval
 
                         # Build package
                         executeTest(
-                            "docker exec -i test-build rpmbuild --define 'pgmajorversion %{nil}' -v -bb --clean" .
-                                " root/rpmbuild/SPECS/pgbackrest.spec",
+                            ($strVm ne VM_NONE ? "docker exec -i test-build " : '') .
+                            "rpmbuild --define 'pgmajorversion %{nil}' -v -bb --clean root/rpmbuild/SPECS/pgbackrest.spec",
                             {bSuppressStdErr => true});
 
                         # Remove build container
-                        executeTest("docker rm -f test-build");
+                        if ($strVm ne VM_NONE)
+                        {
+                            executeTest("docker rm -f test-build");
+                        }
                     }
                 }
 
@@ -1301,7 +1306,7 @@ eval
                 if (!defined($$oyProcess[$iVmIdx]) && $iTestIdx < @{$oyTestRun})
                 {
                     my $oJob = new pgBackRestTest::Common::JobTest(
-                        $oStorageTest, $strBackRestBase, $strTestPath, $strCoveragePath, $$oyTestRun[$iTestIdx], $bDryRun, $bVmOut,
+                        $oStorageTest, $strBackRestBase, $strTestPath, $$oyTestRun[$iTestIdx], $bDryRun, $strVmHost, $bVmOut,
                         $iVmIdx, $iVmMax, $iTestIdx, $iTestMax, $strLogLevel, $strLogLevelTest, $bLogForce, $bShowOutputAsync,
                         $bNoCleanup, $iRetry, !$bNoValgrind, !$bNoCoverage, $bCoverageSummary, !$bNoOptimize, $bBackTrace,
                         $bProfile, $iScale, !$bNoDebug, $bDebugTestTrace, $iBuildMax / $iVmMax < 1 ? 1 : int($iBuildMax / $iVmMax));
@@ -1484,6 +1489,11 @@ eval
     ################################################################################################################################
     # Runs tests
     ################################################################################################################################
+    push(
+        @INC,
+        "${strBackRestBase}/test/.vagrant/bin/" .
+            ((testDefModuleTest($stryModule[0], $stryModuleTest[0]))->{&TESTDEF_CONTAINER} ? $strVm : $strVmHost) .
+            "/libc/blib/arch");
     buildLoadLibC();
 
     my $iRun = 0;
@@ -1502,13 +1512,13 @@ eval
         $strPgVersion ne 'minimal' ? $strPgVersion: undef,          # Pg version
         $stryModule[0], $stryModuleTest[0], \@iyModuleTestRun,      # Module info
         $bVmOut, $bDryRun, $bNoCleanup, $bLogForce,                 # Test options
-        TEST_USER, BACKREST_USER, TEST_GROUP);                      # User/group info
+        TEST_USER, TEST_GROUP);                                     # User/group info
 
     if (!$bNoCleanup)
     {
         if ($oHostGroup->removeAll() > 0)
         {
-            executeTest("sudo rm -rf ${strTestPath}");
+            executeTest("rm -rf ${strTestPath}");
         }
     }
